Very minor

+++ std_e/data_structure/heterogenous_vector.hpp
@@ -32,7 +32,7 @@ class hvector {
   // size
     //! \brief Heterogenous size: number of different types
     static constexpr auto
-    hsize() -> size_t {
+    hsize() -> int {
       return sizeof...(Ts);
     }
